server/pyproject.toml
[project]
name = "pi-security-camera"
version = "0.1.0"
description = "Project about a remote Raspberry Pi security camera"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.119.0",
    "psycopg2-binary>=2.9.11",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.17",
    "mypy>=1.18.2",
    "pytest>=9.0.2",
    "ruff>=0.14.1",
    "types-aiofiles>=25.1.0.20251011",
]

[tool.mypy]
strict = true
disallow_untyped_globals = true
strict_optional = true
